requirements.txt
ccxt==3.1.15
gymnasium==0.29.1
numpy==1.24.3
pandas==2.0.1
torch==2.0.1
matplotlib==3.7.1
scikit-learn==1.2.2
stable-baselines3==2.1.0
python-dotenv==1.0.0
ta-lib==0.4.20
pytorch-lightning==2.0.4
optuna==3.1.1
wandb==0.15.3